# [Bugfix][V1] Exclude HBM used by other processes when calculating peak memory during profile runs