--- a/vllm/v1/worker/gpu_worker.py
+++ b/vllm/v1/worker/gpu_worker.py
@@ -134,7 +134,7 @@ class Worker(WorkerBase):
 
     @torch.inference_mode()
     def determine_available_memory(self) -> int:
-        """Profiles the peak memory usage of the model to determine how much 
+        """Profiles the peak memory usage of the model to determine how much
         memory can be used for KV cache without OOMs.
 
         The engine will first conduct a profiling of the existing memory usage.
@@ -148,7 +148,10 @@ class Worker(WorkerBase):
         torch.cuda.empty_cache()
         torch.cuda.reset_peak_memory_stats()
 
-        _, total_gpu_memory = torch.cuda.mem_get_info()
+        free_gpu_memory_before, total_gpu_memory = torch.cuda.mem_get_info()
+        non_torch_allocations_before = (total_gpu_memory -
+                                        free_gpu_memory_before -
+                                        torch.cuda.memory_allocated())
         # Execute a forward pass with dummy inputs to profile the memory usage
         # of the model.
         self.model_runner.profile_run()
@@ -171,11 +174,16 @@ class Worker(WorkerBase):
         torch.cuda.empty_cache()
         torch_allocated_bytes = torch.cuda.memory_stats(
         )["allocated_bytes.all.current"]
-        total_allocated_bytes = torch.cuda.mem_get_info(
-        )[1] - torch.cuda.mem_get_info()[0]
+        total_allocated_bytes = (torch.cuda.mem_get_info()[1] -
+                                 torch.cuda.mem_get_info()[0])
         non_torch_allocations = total_allocated_bytes - torch_allocated_bytes
-        if non_torch_allocations > 0:
-            peak_memory += non_torch_allocations
+        non_torch_allocations_increase = non_torch_allocations - max(
+            0, non_torch_allocations_before)
+        logger.warning(
+            f"{non_torch_allocations=}, {non_torch_allocations_before=}")
+        logger.warning(f"{non_torch_allocations_increase=}")
+        if non_torch_allocations_increase > 0:
+            peak_memory += non_torch_allocations_increase
         available_kv_cache_memory = (
             total_gpu_memory * self.cache_config.gpu_memory_utilization -
             peak_memory)