@youkaichao 2025-03-07
we should port the memory profiling logic from v0 to v1. i have a complete solution for this. it considers the memory used by other processes.

for reference, please check https://github.com/vllm-project/vllm/blob/70da0c07483bb0cbcc80a2094cf41b64cb30ac0f/vllm/utils.py#L2006