@yeqcharlotte 2025-03-18
close this one in favor of profiler integration